requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
python-decouple>=3.8
psycopg2-binary>=2.9.9

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0

# OAuth
django-allauth>=0.57.0
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=41.0.0
